# handler/user/update.go
package user

import (
	"strconv"

	. "apiserver/handler"
	"apiserver/model"
	"apiserver/pkg/errno"
	"apiserver/util"

	"github.com/gin-gonic/gin"
	"github.com/gocoder2009/log-for-apiserver"
	"github.com/gocoder2009/log-for-apiserver/lager"
)

// Update godoc
//
//	@Summary		Update a user info by the user identifier
//	@Description	Update a user by ID
//	@Tags			user
//	@Accept			json
//	@Produce		json
//	@Param			id		path		uint64				true	"The user's database id index num"
//	@Param			user	body		model.UserModel		true	"The user info"
//	@Success		200		{object}	handler.Response	"{"code":0,"message":"OK","data":null}"
//	@Router			/v1/user/{id} [put]
func Update(c *gin.Context) {
	log.Info("Update function called.", lager.Data{"X-Request-Id": util.GetReqID(c)})
	// Get the user id from the url parameter.
	userId, _ := strconv.Atoi(c.Param("id"))

	// Binding the user data.
	var u model.UserModel
	if err := c.Bind(&u); err != nil {
		SendResponse(c, errno.ErrBind, nil)
		return
	}

	// We update the record based on the user id.
	u.Id = uint64(userId)

	// Validate the data.
	if err := u.Validate(); err != nil {
		SendResponse(c, errno.ErrValidation, nil)
		return
	}

	// Encrypt the user password.
	if err := u.Encrypt(); err != nil {
		SendResponse(c, errno.ErrEncrypt, nil)
		return
	}

	// Save changed fields.
	if err := u.Update(); err != nil {
		SendResponse(c, errno.ErrDatabase, nil)
		return
	}

	SendResponse(c, nil, nil)
}